[src] Move utils files into utils folder
